add chapter handling code
All chapters added between segments by the edl specification are removed

In addition all chapters that are within the merge-threshold for ordered
chapters are also removed to keep things tidy.

The matroska specification does not mention a need for chapters between
files, and vlc does not create any either.
Therefore, we'll leave chapters entirely up to the encoders